requirements-dev.txt
-r requirements.txt

# Development tools
black==23.12.1
isort==5.13.2
flake8==7.0.0
flake8-docstrings==1.7.0
mypy==1.8.0
pre-commit==3.6.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Development utilities
watchfiles==0.21.0
ipdb==0.13.13
python-dotenv==1.0.0